feat: remove old vehicle status widget
